feat: fetch template list on mount
